backend/requirements.txt
fastapi
fastapi-users[sqlalchemy]==13.0.0
pydantic>=2.0,<3.0
pydantic-settings
email-validator
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt
uvicorn[standard]
granian
jinja2>=3.1.3,<4.0
python-multipart
aiohttp
e2b==1.3.4rc1
docker>=7.1.0
claude-agent-sdk>=0.1.14
mcp
sqladmin[full]
httpx
aiosmtplib
prometheus-fastapi-instrumentator
slowapi
celery[redis]
sse-starlette
redis
tenacity==8.2.3
PyYAML>=6.0
python-json-logger>=2.0.0
mypy>=1.8.0
types-redis
types-PyYAML
